refactor: replace handler_auth with handler_oauth
- Delete handler_auth.go with inline OAuth logic
- Add handler_oauth.go delegating to oauth.Service
- Simplify Login and Callback handlers
- Remove session and OAuth config dependencies
